Remove IO async for Room takes care background thread for transaction automatically

+++ Acornote_Kotlin/app/src/main/java/tonnysunm/com/acornote/ui/detail/DetailViewModel.kt
@@ -9,14 +9,16 @@ import tonnysunm.com.acornote.model.Item
 import tonnysunm.com.acornote.model.Repository
 
 
-class DetailViewModelFactory (private val application: Application, private val folderId: Int) : ViewModelProvider.NewInstanceFactory() {
+class DetailViewModelFactory(private val application: Application, private val folderId: Long) :
+    ViewModelProvider.NewInstanceFactory() {
 
     @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>) = DetailViewModel(application, folderId) as T
+    override fun <T : ViewModel> create(modelClass: Class<T>) =
+        DetailViewModel(application, folderId) as T
 
 }
 
-class DetailViewModel(application: Application, folderId: Int) : AndroidViewModel(application) {
+class DetailViewModel(application: Application, folderId: Long) : AndroidViewModel(application) {
 
     private val repository: Repository by lazy { Repository(application) }
